Caster controller 2.0.0 featuring a bunch of new features:
Pin Control is now a drawer, and has an easy access button on the main screen.
Forward/Backward/Beginning buttons are always enabled once a ribbon is started.
Added Start Pump/Stop Pump and Deliver line buttons.
Added new preference that enables automatic stopping/starting of the pump when the ribbon is
started and stopped.
Added "Lines in Job" drawer that shows the lines in the job, with the current line highlighted.